Perform principal coordinate analysis (PCoA).

Perform principal coordinate analysis (PCoA).

Usage

Microbiome.PCoA(
  data,
  sample,
  method = "bray",
  x = "PCo1",
  y = "PCo2",
  size = 2,
  color = "group",
  alpha = 1
)

Arguments

data

Feature table, with rows representing samples and columns representing feature values, such as OTUs.

sample

Sample table, with the first column containing sample names. There are no specific requirements for the names of the subsequent columns, but the sample names must match those in the feature table.

method

The method for calculating distances in vegan::vegdist() is defaulted to Bray-Curtis.

x

The principal coordinate used for the X-axis in the plot is defaulted to PCo1.

y

The principal coordinate used for the y-axis in the plot is defaulted to PCo2.

size

The size of point.

color

Column names in the sample data frame used for coloring the points.

alpha

The alpha of point.

Value

A list containing a data frame and a ggplot2 object.

Examples

library(dplyr)
library(biohelpers)

data(df.pcoa.otu)
data(df.pcoa.sample)

Microbiome.PCoA(data = df.pcoa.otu, sample = df.pcoa.sample) -> pcoa.res
#> Joining with `by = join_by(sample)`