Calculate the top ten microorganisms by abundance at different taxonomic levels.

Calculate the top ten microorganisms by abundance at different taxonomic levels.

Usage

Microbiome.top_10(
  data,
  sample,
  taxo,
  which = "phylum",
  by = "sum",
  method = "wilcox",
  ref = "CK",
  pvalue = 0.05
)

Arguments

data

Feature table, such as OTU abundance information, where rows represent samples and columns represent feature values.

sample

Sample table, where the first column is the sample name, the second column and onwards contain grouping information, and there must be a column named 'group'.

taxo

Classification information, such as the classification information of OTUs.

which

Which classification information to analyze and plot, with 'phylum' as the default.

by

Calculate the Top 10 by summation or averaging, with 'sum' as the default.

method

Statistical analysis method, with 'wilcox' (Wilcox test) as the default.

ref

The name of the control group used for statistical analysis, with 'CK' as the default.

pvalue

Threshold for the p-value, with a default of 0.05.

Value

A list containing a data frame and a ggplot2 object.

Examples

library(dplyr)
library(biohelpers)

data(df.top10.otu)
data(df.top10.sample)
data(df.top10.class)

Microbiome.top_10(df.top10.otu, df.top10.sample, df.top10.class) -> top10.res
#> Joining with `by = join_by(sample)`
#> Joining with `by = join_by(OTU)`
#> Joining with `by = join_by(sample)`
#> Joining with `by = join_by(taxo)`