Perform GO enrichment analysis of differentially expressed genes.

Perform GO enrichment analysis of differentially expressed genes.

Usage

RNASeq.enrich_GO(gene, go.db, pAdjustMethod = "BH", p.adjust = 0.05)

Arguments

gene

The names of differentially expressed genes are a vector.

go.db

The GO database used for GO enrichment analysis needs to be in data frame format and must contain the following columns: gene, go.id, go.term. The go.ontology column is optional.

pAdjustMethod

Methods for adjusting p-values can be selected from "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", and "none", with "BH" as the default.

p.adjust

Threshold for the adjusted P value.

Value

A data frame.

Examples

library(biohelpers)

data(df.rnaseq.go)
data(df.rnaseq.degs)

RNASeq.enrich_GO(gene = df.rnaseq.degs$gene, go.db = df.rnaseq.go) -> go.res