Identifying differential abundance microbiomes using LEfSe.

Identifying differential abundance microbiomes using LEfSe.

Usage

call_DAMs_LEfSe(
  data,
  sample,
  groupCol = "group",
  kruskal.threshold = 0.05,
  wilcox.threshold = 0.05,
  lda.threshold = 1
)

Arguments

data

Feature table, with rows representing feature values and columns representing sample names.

sample

Sample table, with rows representing sample name and columns representing group information.

groupCol

The column name where the grouping information is located, default is 'group'.

kruskal.threshold

The threshold for the Kruskal test, default is 0.05.

wilcox.threshold

The threshold for the Wilcox test, default is 0.05.

lda.threshold

The threshold for the LDA test, default is 1.

Value

A data frame.

Examples

library(biohelpers)
library(dplyr)

data(df.call_DAMs_LEfSe.otu)
data(df.call_DAMs_LEfSe.sample)

call_DAMs_LEfSe(df.call_DAMs_LEfSe.otu, df.call_DAMs_LEfSe.sample) -> lefse.result
#> Warning: The 'groupCol' argument is deprecated, use 'classCol' instead.
#> Warning: Convert counts to relative abundances with 'relativeAb()'
#> The outcome variable is specified as 'group' and the reference category is 'CK'.
#>  See `?factor` or `?relevel` to change the reference category.
#> Warning: Variables in the input are collinear. Try only with the terminal nodes using `get_terminal_nodes` function