Identifying Differentially Expressed Genes Using DESeq2

Identifying Differentially Expressed Genes Using DESeq2

Usage

call_DEGs_DESeq2(data, sample, group, log2FoldChange = 1, padj = 0.05)

Arguments

data

Expression matrix must be integers, with rows representing gene names and columns representing sample names.

sample

Sample information table, with rows representing sample names.

group

Grouping information of samples, the column names in the @param sample table, must be factors.

log2FoldChange

The threshold for the absolute value of log2FoldChange is set to 1 by default.

padj

The threshold for the adjusted p-value is set to 0.05 by default.

Value

A data frame.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(biohelpers)

data(df.gene.rnaseq)
data(df.sample.rnaseq)
call_DEGs_DESeq2(data = df.gene.rnaseq, 
                 sample = df.sample.rnaseq, 
                 group = "group") -> degs
#> converting counts to integer mode
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing