Title

Title

Usage

reorder2heatmap(data, col, row, value)

Arguments

data

Input data in a data frame format, specifically in long data format, tidyr format.

col

Typically, the column name where the sample names are located.

row

Typically, the column name where the feature names are located.

value

The column name where the numerical values needed for plotting the heatmap are located.

Value

A data frame.

Examples

library(dplyr)
library(biohelpers)

data(df.reorder2heatmap)

reorder2heatmap(
  data = df.reorder2heatmap,
  col = "sample",
  row = "meta",
  value = "value"
) -> df.reorder